pyproject.toml
[project]
name = "finkit"
description = "financial kit"
version = "0.2.2"
dependencies = [
    "pyyaml >= 6.0",
    "numpy >= 1.23",
    "scipy >= 1.9",
    "pandas >= 1.5",
    "beautifulsoup4 >= 4.11",
    "lxml >= 4.9",
    "crontab >= 0.23",
    "sqlalchemy >=1.4,<2.0",
    "chinesecalendar >= 1.8",
    "pymssql == 2.2.7",
    "toml == 0.10.2",
    "akshare >= 1.8"
]

[project.scripts]
finkit = "finkit.__main__:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
#include = ["finkit"]

[tool.setuptools.package-data]
finkit = ["*.csv"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
